requirements.txt
certifi
requests
pyyaml
mongoengine
netaddr
elasticsearch
